avoid multiple watcher threads
The ReadDirectoryChanges backend creates a thread for every directory
being watched. That causes the events for multiplie directories coming
in asynchronously and that makes testing with the current test structure
impossible.  Therefore, just one directory should be watched, and
another test that has to be written must check that multiple directories
can be watched simultaneously but with a test that does not rely on a
strict order.